admin dashboard change

+++ controller/adminController.js
@@ -5,18 +5,25 @@ const Doctor = require('../model/doctor.model');
 
 router.get('/doctor', (req, res) => {
     if (req.session.doctor) {
-        Doctor.find((err, docs) => {
-            if (!err) {
-                res.render('doctor/adminDashboard.hbs', {
-                    titel: "Doctors Page",
-                    style: 'adminDashboard.css',
-                    doctors: docs,
-                    admin: "Admin "
-                });
-            } else {
-                console.log('Error during record insertion : ' + err);
-            }
-        });
+        if (req.session.doctor.userName == "admindoctor") {
+            Doctor.find((err, docs) => {
+                if (!err) {
+                    docs = docs.filter((doctor) => {
+                        return doctor.userName != "admindoctor"
+                    })
+                    res.render('doctor/adminDashboard.hbs', {
+                        titel: "Doctors Page",
+                        style: 'adminDashboard.css',
+                        doctors: docs,
+                        admin: "Admin "
+                    });
+                } else {
+                    console.log('Error during record insertion : ' + err);
+                }
+            });
+        } else {
+            res.redirect('/login');
+        }
     } else {
         res.redirect('/login');
     }
@@ -24,19 +31,23 @@ router.get('/doctor', (req, res) => {
 
 router.get('/update/:id', (req, res) => {
     if (req.session.doctor) {
-        Doctor.findById(req.params.id, (err, doc) => {
-            if (!err) {
-                res.render('doctor/addDoctor.hbs', {
-                    titel: "Doctors Page",
-                    style: 'addDoctor.css',
-                    doctorTitle: 'Update',
-                    doctor: doc,
-                    admin: "Admin "
-                });
-            } else {
-                console.log('Error during record insertion : ' + err);
-            }
-        });
+        if (req.session.doctor.userName == "admindoctor") {
+            Doctor.findById(req.params.id, (err, doc) => {
+                if (!err) {
+                    res.render('doctor/addDoctor.hbs', {
+                        titel: "Doctors Page",
+                        style: 'addDoctor.css',
+                        doctorTitle: 'Update',
+                        doctor: doc,
+                        admin: "Admin "
+                    });
+                } else {
+                    console.log('Error during record insertion : ' + err);
+                }
+            });
+        } else {
+            res.redirect('/login');
+        }
     } else {
         res.redirect('/login');
     }
@@ -45,13 +56,17 @@ router.get('/update/:id', (req, res) => {
 
 router.get('/delete/:id', (req, res) => {
     if (req.session.doctor) {
-        Doctor.findByIdAndRemove(req.params.id, (err, doc) => {
-            if (!err) {
-                res.redirect('/admin/doctor');
-            } else {
-                console.log('Error during record insertion : ' + err);
-            }
-        });
+        if (req.session.doctor.userName == "admindoctor") {
+            Doctor.findByIdAndRemove(req.params.id, (err, doc) => {
+                if (!err) {
+                    res.redirect('/admin/doctor');
+                } else {
+                    console.log('Error during record insertion : ' + err);
+                }
+            });
+        } else {
+            res.redirect('/login');
+        }
     } else {
         res.redirect('/login');
     }
@@ -59,11 +74,15 @@ router.get('/delete/:id', (req, res) => {
 
 router.get('/', (req, res) => {
     if (req.session.doctor) {
-        res.render('doctor/home.hbs', {
-            titel: "Home",
-            style: 'home.css',
-            admin: "Admin "
-        });
+        if (req.session.doctor.userName == "admindoctor") {
+            res.render('doctor/home.hbs', {
+                titel: "Home",
+                style: 'home.css',
+                admin: "Admin "
+            });
+        } else {
+            res.redirect('/login');
+        }
     } else {
         res.redirect('/login');
     }
@@ -71,12 +90,16 @@ router.get('/', (req, res) => {
 
 router.get('/add', (req, res) => {
     if (req.session.doctor) {
-        res.render('doctor/addDoctor.hbs', {
-            titel: "Add Doctor",
-            style: 'addDoctor.css',
-            doctorTitle: 'Add New',
-            admin: "Admin "
-        });
+        if (req.session.doctor.userName == "admindoctor") {
+            res.render('doctor/addDoctor.hbs', {
+                titel: "Add Doctor",
+                style: 'addDoctor.css',
+                doctorTitle: 'Add New',
+                admin: "Admin "
+            });
+        } else {
+            res.redirect('/login');
+        }
     } else {
         res.redirect('/login');
     }
@@ -84,10 +107,14 @@ router.get('/add', (req, res) => {
 
 router.post('/add', (req, res) => {
     if (req.session.doctor) {
-        if (req.body._id == '') {
-            insertDoctor(req, res);
+        if (req.session.doctor.userName == "admindoctor") {
+            if (req.body._id == '') {
+                insertDoctor(req, res);
+            } else {
+                updateInfo(req, res);
+            }
         } else {
-            updateInfo(req, res);
+            res.redirect('/login');
         }
     } else {
         res.redirect('/login');
